pyproject.toml
[project]
name = "langchain-chatbot"
version = "0.1.0"
description = "LangChain Chatbot with FastAPI"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langchain>=0.1.20",
    "langchain-community>=0.0.10",
    "langchain-classic>=0.0.7",
    "langchain-core>=0.1.10",
    "zhipuai>=2.0.1",
    "tavily-python>=0.3.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
start = "backend.main:app"
dev = "backend.main:app --reload --host 127.0.0.1 --port 8000"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (real API calls)",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
]
